backend/requirements.txt
Flask==3.0.0
# TODO: Is this needed?
psycopg2_binary==2.9.9
Flask_Migrate==4.0.5
flask_sqlalchemy==3.1.1
marshmallow==3.20.1
requests==2.31.0
# TODO: Try to avoid using CORS
Flask-Cors==4.0.0
pytest==7.4.3
pytest-mock==3.12.0
SQLAlchemy==2.0.23
freezegun==1.2.2
alembic==1.12.1
python-dotenv==1.0.0